moved dependency package to src dir
