--- .idea/dictionaries/maverick.xml
@@ -1,9 +1,0 @@
-<component name="ProjectDictionaryState">
-  <dictionary name="maverick">
-    <words>
-      <w>dataset</w>
-      <w>preprocess</w>
-      <w>preprocessing</w>
-    </words>
-  </dictionary>
-</component>
