--- .idea/retail-analytics.iml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module type="PYTHON_MODULE" version="4">
-  <component name="NewModuleRootManager">
-    <content url="file://$MODULE_DIR$" />
-    <content url="jar://$USER_HOME$/program-files/spark/spark-2.4.4-bin-hadoop2.7/python/lib/py4j-0.10.7-src.zip!/" />
-    <content url="jar://$USER_HOME$/program-files/spark/spark-2.4.4-bin-hadoop2.7/python/lib/pyspark.zip!/" />
-    <orderEntry type="jdk" jdkName="Python 3.7 (main)" jdkType="Python SDK" />
-    <orderEntry type="sourceFolder" forTests="false" />
-  </component>
-  <component name="TestRunnerService">
-    <option name="PROJECT_TEST_RUNNER" value="Twisted Trial" />
-  </component>
-</module>
