--- README.md
@@ -1,17 +1,0 @@
-**RUN below command to submit the spark job:**
-
-steps to submit the spark job
-
-**STEP 1:** build build_dependencies.sh
-
-**STEP 2:** spark-submit --master local[1] --py-files packages.zip --files config/preprocess_config.json main.py 
-
-TODO:
-Overview of folder structure and setup
-Documentation on Each Job
-Documentation on Notebooks
-Documentation on Test Cases
-Reference Links
-
-
-
