--- assets/input-data/info.txt
@@ -1,1 +1,0 @@
-input data to be processed goes here
