--- config/preprocess_config.json
@@ -1,3 +1,0 @@
-{
-  "steps_per_floor": 21
-}
