--- config/spark/spark-defaults.conf
@@ -1,26 +1,0 @@
-spark.hadoop.fs.s3a.impl org.apache.hadoop.fs.s3a.S3AFileSystem
-spark.hadoop.fs.s3a.access.key test
-spark.hadoop.fs.s3a.secret.key test
-spark.hadoop.fs.s3a.endpoint.region us-east-1
-spark.hadoop.fs.s3a.endpoint http://localstack:4566
-spark.hadoop.fs.s3a.path.style.access true
-spark.sql.extensions org.apache.iceberg.spark.extensions.IcebergSparkSessionExtensions
-spark.sql.catalog.data_catalog org.apache.iceberg.spark.SparkCatalog
-spark.sql.catalog.data_catalog.catalog-impl org.apache.iceberg.jdbc.JdbcCatalog
-spark.sql.catalog.data_catalog.warehouse s3a://my-bucket/iceberg/data/prefix/
-spark.sql.catalog.data_catalog.uri jdbc:postgresql://postgres:5432/iceberg_catalog
-spark.sql.catalog.data_catalog.jdbc.verifyServerCertificate true
-spark.sql.catalog.data_catalog.jdbc.useSSL true
-spark.sql.catalog.data_catalog.jdbc.user de_user
-spark.sql.catalog.data_catalog.jdbc.password de_pwd
-spark.jars.packages org.apache.iceberg:iceberg-spark-runtime-3.5_2.12:1.4.3,org.postgresql:postgresql:42.7.2,com.amazonaws:aws-java-sdk-bundle:1.12.262,org.apache.hadoop:hadoop-aws:3.3.4
-spark.jars.ivy /opt/spark/ivy/
-
-
-
---conf spark.jars.packages=org.apache.iceberg:iceberg-spark-runtime-3.3_2.12:1.4.3,com.amazonaws:aws-java-sdk-bundle:1.12.262,org.apache.hadoop:hadoop-aws:3.3.4
-
-
---extra-jars s3://amazonapi-test-bucket02/pii_918_jars/bundle-2.24.8.jar,s3://amazonapi-test-bucket02/pii_918_jars/iceberg-spark-runtime-3.3_2.12-1.4.3.jar
-
-s3://amazonapi-test-bucket02/pii_918_jars/org.apache.hadoop_hadoop-aws-3.3.4.jar
