--- derby.log
@@ -1,13 +1,0 @@
-----------------------------------------------------------------
-Wed Sep 23 16:57:13 IST 2020:
-Booting Derby version The Apache Software Foundation - Apache Derby - 10.12.1.1 - (1704137): instance a816c00e-0174-bab7-a8b3-00000a9d7b70 
-on database directory /home/maverick/workspace/personal-workspace/Data-engineerng-use-cases/retail-analytics/metastore_db with class loader org.apache.spark.sql.hive.client.IsolatedClientLoader$$anon$1@4c06451a 
-Loaded from file:/home/maverick/program-files/spark/spark-2.4.4-bin-hadoop2.7/jars/derby-10.12.1.1.jar
-java.vendor=Private Build
-java.runtime.version=1.8.0_265-8u265-b01-0ubuntu2~20.04-b01
-user.dir=/home/maverick/workspace/personal-workspace/Data-engineerng-use-cases/retail-analytics
-os.name=Linux
-os.arch=amd64
-os.version=5.4.0-42-generic
-derby.system.home=null
-Database Class Loader started - derby.database.classpath=''
