--- main.tf
@@ -1,41 +1,0 @@
-provider "aws" {
-  access_key                  = "localstack_access_key"
-  secret_key                  = "localstack_secret_key"
-  region                      = "us-east-1"
-
-  s3_use_path_style           = true
-  skip_credentials_validation = true
-  skip_metadata_api_check     = true
-  skip_requesting_account_id  = true
-
-  endpoints {
-    apigateway     = "http://localstack:4566"
-    apigatewayv2   = "http://localstack:4566"
-    cloudformation = "http://localstack:4566"
-    cloudwatch     = "http://localstack:4566"
-    dynamodb       = "http://localstack:4566"
-    ec2            = "http://localstack:4566"
-    es             = "http://localstack:4566"
-    elasticache    = "http://localstack:4566"
-    firehose       = "http://localstack:4566"
-    iam            = "http://localstack:4566"
-    kinesis        = "http://localstack:4566"
-    lambda         = "http://localstack:4566"
-    rds            = "http://localstack:4566"
-    redshift       = "http://localstack:4566"
-    route53        = "http://localstack:4566"
-    s3             = "http://localstack:4566"
-    secretsmanager = "http://localstack:4566"
-    ses            = "http://localstack:4566"
-    sns            = "http://localstack:4566"
-    sqs            = "http://localstack:4566"
-    ssm            = "http://localstack:4566"
-    stepfunctions  = "http://localstack:4566"
-    sts            = "http://localstack:4566"
-  }
-}
-
-
-resource "aws_s3_bucket" "test-bucket" {
-  bucket = "my-bucket"
-}
