--- pytest.ini
@@ -1,13 +1,0 @@
-# pytest.ini
-#to run the tests RUN:
-# python -m pytest -v / pytest -v / py.test -v  :  -v for verbose data
-# python -k  <test ame>  :  -k run all tests with specific string in test name
-#  pytest -rxs:  -rxs print reasons for skipping tests if any
-# pytest -m "dbtests" --capture=no  show all content with marker "dbtests"; --capture=no is used to print output to console test"
-
-[pytest]
-markers =
-    job1: marks tests of job1 (deselect with '-m "not job1"').
-    job_0: marks tests of job_0 (deselect with '-m "not job_0"').
-
-spark_home = /home/maverick/program-files/spark/spark-2.4.4-bin-hadoop2.7
